feat: Harden FSRS-7 numerics against pinned fsrs-rs f32 reference

+++ backend/KotobaApi/Srs/Domain/Fsrs7MemoryState.cs
@@ -2,8 +2,9 @@ namespace KotobaApi.Srs.Domain;
 
 /// <summary>
 /// FSRS-7 dual-trace memory state. Stability values are expressed in days.
+/// Single-precision floats mirror the Rust f32 runtime of fsrs-rs.
 /// </summary>
 public readonly record struct Fsrs7MemoryState(
-    double Stability,
-    double FastStability,
-    double Difficulty);
+    float Stability,
+    float FastStability,
+    float Difficulty);
